third_party/blink/web_tests/platform/generic/http/tests/websocket/unpaired-surrogates-in-close-reason-expected.txt

Checks whether unpaired surrogates in a close reason are silently converted to U+FFFD and SYNTAX_ERR is not thrown.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected.
PASS closeEvent.wasClean is true
PASS successfullyParsed is true

TEST COMPLETE
